backend/requirements.txt
anthropic[bedrock]>=0.25.0
boto3>=1.28.0
langgraph>=0.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
